Split test files, add gas payment / synthetic coin tests
- Split coin-flows.test.ts into 3 files:
  - coin-flows.test.ts: original base tests (12 tests)
  - coin-flows-framework.test.ts: framework MoveCall + chain + transfer-to-self (27 tests)
  - coin-flows-intents.test.ts: tx.coin/tx.balance integration + gas payment (18 tests)
- Add 4 new gas payment tests: empty payment, synthetic coin digest
  parsing, synthetic + real coin combined, splits with synthetic coins
- Handle synthetic coin reservation refs in gasCoins analyzer
- Exclude synthetic coin refs from object fetching
- Copy coin-reservation.ts utility from @mysten/sui (not yet exported)

68 tests total, all passing.
