+++ packages/wallet-sdk/test/transaction-analyzer/coin-flows.test.ts
@@ -3,7 +3,7 @@
 
 import { describe, it, expect } from 'vitest';
 import { Transaction } from '@mysten/sui/transactions';
-import { normalizeStructTag } from '@mysten/sui/utils';
+
 import { analyze } from '../../src/transaction-analyzer/analyzer.js';
 import { coinFlows } from '../../src/transaction-analyzer/rules/coin-flows.js';
 import { MockSuiClient } from '../mocks/MockSuiClient.js';
@@ -15,9 +15,6 @@ import {
 	TEST_USDC_COIN_ID,
 	TEST_WETH_COIN_ID,
 } from '../mocks/mockData.js';
-
-const SUI = '0x0000000000000000000000000000000000000000000000000000000000000002::sui::SUI';
-const USDC = '0x0000000000000000000000000000000000000000000000000000000000000a0b::usdc::USDC';
 
 describe('TransactionAnalyzer - Coin Flows Rule', () => {
 	it('should handle empty transactions with no coin flows', async () => {
@@ -442,1120 +439,5 @@ describe('TransactionAnalyzer - Coin Flows Rule', () => {
 				'0x0000000000000000000000000000000000000000000000000000000000000b0c::weth::WETH',
 		);
 		expect(wethFlow?.amount).toBe(2500000000000000000n); // WETH balance consumed
-	});
-});
-
-// Helper: build a transaction with a FundsWithdrawal input and a redeem MoveCall.
-function buildWithdrawalTx(opts: {
-	sender?: string;
-	withdrawalAmount: bigint;
-	coinType: string;
-	redeemTarget: '0x2::balance::redeem_funds' | '0x2::coin::redeem_funds';
-}) {
-	const tx = new Transaction();
-	tx.setSender(opts.sender ?? DEFAULT_SENDER);
-
-	const withdrawalInput = tx.withdrawal({
-		amount: opts.withdrawalAmount,
-		type: opts.coinType,
-	});
-
-	const result = tx.moveCall({
-		target: opts.redeemTarget,
-		typeArguments: [opts.coinType],
-		arguments: [withdrawalInput],
-	});
-
-	return { tx, result };
-}
-
-describe('Coin Flows - Framework MoveCall Tests', () => {
-	// --- Address balance operations ---
-
-	it('tracks coin::redeem_funds as an owned coin outflow', async () => {
-		const client = new MockSuiClient();
-		const { tx, result } = buildWithdrawalTx({
-			withdrawalAmount: 500000000n,
-			coinType: '0xa0b::usdc::USDC',
-			redeemTarget: '0x2::coin::redeem_funds',
-		});
-
-		tx.transferObjects([result], tx.pure.address('0x456'));
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow?.amount).toBe(500000000n);
-	});
-
-	it('tracks balance::redeem_funds as an owned balance outflow', async () => {
-		const client = new MockSuiClient();
-		const { tx, result } = buildWithdrawalTx({
-			withdrawalAmount: 300000000n,
-			coinType: '0xa0b::usdc::USDC',
-			redeemTarget: '0x2::balance::redeem_funds',
-		});
-
-		// Consume the balance in a generic MoveCall
-		tx.moveCall({
-			target: '0x999::test::consume_coin',
-			typeArguments: [normalizeStructTag('0xa0b::usdc::USDC')],
-			arguments: [result, result], // second arg doesn't matter, consume_coin takes 2
-		});
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow?.amount).toBe(300000000n);
-	});
-
-	it('does not count sponsored withdrawal as sender outflow', async () => {
-		const client = new MockSuiClient();
-
-		// Build a simple tx with a Sender withdrawal, then patch the JSON to use Sponsor
-		const { tx } = buildWithdrawalTx({
-			withdrawalAmount: 500000000n,
-			coinType: '0xa0b::usdc::USDC',
-			redeemTarget: '0x2::coin::redeem_funds',
-		});
-
-		tx.transferObjects(
-			[
-				tx.moveCall({
-					target: '0x2::coin::redeem_funds',
-					typeArguments: ['0xa0b::usdc::USDC'],
-					arguments: [tx.withdrawal({ amount: 500000000n, type: '0xa0b::usdc::USDC' })],
-				}),
-			],
-			tx.pure.address('0x456'),
-		);
-
-		// Get JSON and patch withdrawFrom to Sponsor
-		const json = JSON.parse(await tx.toJSON());
-		for (const input of json.inputs) {
-			if (input.FundsWithdrawal) {
-				input.FundsWithdrawal.withdrawFrom = { Sponsor: true };
-			}
-		}
-
-		const results = await analyze({ coinFlows }, { client, transaction: JSON.stringify(json) });
-
-		// Sponsored withdrawal should not count as sender outflow
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow).toBeUndefined();
-	});
-
-	it('handles coin::send_funds to sender (no outflow)', async () => {
-		const client = new MockSuiClient();
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		const coin = tx.object(TEST_COIN_1_ID);
-		const [splitCoin] = tx.splitCoins(coin, [200000000n]);
-
-		tx.moveCall({
-			target: '0x2::coin::send_funds',
-			typeArguments: ['0x2::sui::SUI'],
-			arguments: [splitCoin, tx.pure.address(DEFAULT_SENDER)],
-		});
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		const suiFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === SUI);
-		expect(suiFlow?.amount).toBe(10000000n); // gas budget only
-	});
-
-	it('handles coin::send_funds to other address (outflow)', async () => {
-		const client = new MockSuiClient();
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		const coin = tx.object(TEST_COIN_1_ID);
-		const [splitCoin] = tx.splitCoins(coin, [200000000n]);
-
-		tx.moveCall({
-			target: '0x2::coin::send_funds',
-			typeArguments: ['0x2::sui::SUI'],
-			arguments: [splitCoin, tx.pure.address('0x456')],
-		});
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		const suiFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === SUI);
-		expect(suiFlow?.amount).toBe(210000000n); // 200M + gas budget
-	});
-
-	it('handles balance::send_funds to sender (no outflow)', async () => {
-		const client = new MockSuiClient();
-		const { tx, result: balResult } = buildWithdrawalTx({
-			withdrawalAmount: 100000000n,
-			coinType: '0xa0b::usdc::USDC',
-			redeemTarget: '0x2::balance::redeem_funds',
-		});
-
-		tx.moveCall({
-			target: '0x2::balance::send_funds',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [balResult, tx.pure.address(DEFAULT_SENDER)],
-		});
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow?.amount).toBe(0n);
-	});
-
-	it('handles balance::send_funds to other address (outflow)', async () => {
-		const client = new MockSuiClient();
-		const { tx, result: balResult } = buildWithdrawalTx({
-			withdrawalAmount: 100000000n,
-			coinType: '0xa0b::usdc::USDC',
-			redeemTarget: '0x2::balance::redeem_funds',
-		});
-
-		tx.moveCall({
-			target: '0x2::balance::send_funds',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [balResult, tx.pure.address('0x456')],
-		});
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow?.amount).toBe(100000000n);
-	});
-
-	// --- Conversions ---
-
-	it('handles coin::into_balance (value preserved, not double-counted)', async () => {
-		const client = new MockSuiClient();
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		const coin = tx.object(TEST_COIN_1_ID); // 5 SUI
-		const [splitCoin] = tx.splitCoins(coin, [300000000n]);
-
-		const balance = tx.moveCall({
-			target: '0x2::coin::into_balance',
-			typeArguments: ['0x2::sui::SUI'],
-			arguments: [splitCoin],
-		});
-
-		// Consume the balance in a generic MoveCall
-		tx.moveCall({
-			target: '0x999::test::consume_coin',
-			typeArguments: [normalizeStructTag('0x2::sui::SUI')],
-			arguments: [balance, balance],
-		});
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		const suiFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === SUI);
-		// 300M split (consumed via into_balance → generic consume) + gas budget
-		expect(suiFlow?.amount).toBe(310000000n);
-	});
-
-	it('handles coin::from_balance conversion', async () => {
-		const client = new MockSuiClient();
-		const { tx, result: balResult } = buildWithdrawalTx({
-			withdrawalAmount: 200000000n,
-			coinType: '0xa0b::usdc::USDC',
-			redeemTarget: '0x2::balance::redeem_funds',
-		});
-
-		const coin = tx.moveCall({
-			target: '0x2::coin::from_balance',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [balResult],
-		});
-
-		tx.transferObjects([coin], tx.pure.address('0x456'));
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow?.amount).toBe(200000000n);
-	});
-
-	it('handles coin::from_balance → transfer to self (no outflow)', async () => {
-		const client = new MockSuiClient();
-		const { tx, result: balResult } = buildWithdrawalTx({
-			withdrawalAmount: 200000000n,
-			coinType: '0xa0b::usdc::USDC',
-			redeemTarget: '0x2::balance::redeem_funds',
-		});
-
-		const coin = tx.moveCall({
-			target: '0x2::coin::from_balance',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [balResult],
-		});
-
-		tx.transferObjects([coin], tx.pure.address(DEFAULT_SENDER));
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow?.amount).toBe(0n);
-	});
-
-	// --- Split operations via MoveCall ---
-
-	it('handles coin::split via MoveCall', async () => {
-		const client = new MockSuiClient();
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		const coin = tx.object(TEST_COIN_1_ID); // 5 SUI
-
-		const splitResult = tx.moveCall({
-			target: '0x2::coin::split',
-			typeArguments: ['0x2::sui::SUI'],
-			arguments: [coin, tx.pure.u64(200000000n)],
-		});
-
-		tx.transferObjects([splitResult], tx.pure.address('0x456'));
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		const suiFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === SUI);
-		expect(suiFlow?.amount).toBe(210000000n); // 200M + gas budget
-	});
-
-	it('handles balance::split via MoveCall', async () => {
-		const client = new MockSuiClient();
-		const { tx, result: balResult } = buildWithdrawalTx({
-			withdrawalAmount: 500000000n,
-			coinType: '0xa0b::usdc::USDC',
-			redeemTarget: '0x2::balance::redeem_funds',
-		});
-
-		const splitBal = tx.moveCall({
-			target: '0x2::balance::split',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [balResult, tx.pure.u64(200000000n)],
-		});
-
-		// Send original remainder (300M) back to sender
-		tx.moveCall({
-			target: '0x2::balance::send_funds',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [balResult, tx.pure.address(DEFAULT_SENDER)],
-		});
-
-		// Consume the split balance
-		tx.moveCall({
-			target: '0x999::test::consume_coin',
-			typeArguments: [normalizeStructTag('0xa0b::usdc::USDC')],
-			arguments: [splitBal, splitBal],
-		});
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow?.amount).toBe(200000000n); // Only split amount, remainder returned
-	});
-
-	it('handles coin::take (Balance → Coin split)', async () => {
-		const client = new MockSuiClient();
-		const { tx, result: balResult } = buildWithdrawalTx({
-			withdrawalAmount: 500000000n,
-			coinType: '0xa0b::usdc::USDC',
-			redeemTarget: '0x2::balance::redeem_funds',
-		});
-
-		const takenCoin = tx.moveCall({
-			target: '0x2::coin::take',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [balResult, tx.pure.u64(150000000n)],
-		});
-
-		tx.transferObjects([takenCoin], tx.pure.address('0x456'));
-
-		// Send remaining balance back to sender
-		tx.moveCall({
-			target: '0x2::balance::send_funds',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [balResult, tx.pure.address(DEFAULT_SENDER)],
-		});
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow?.amount).toBe(150000000n);
-	});
-
-	it('handles balance::withdraw_all', async () => {
-		const client = new MockSuiClient();
-		const { tx, result: balResult } = buildWithdrawalTx({
-			withdrawalAmount: 500000000n,
-			coinType: '0xa0b::usdc::USDC',
-			redeemTarget: '0x2::balance::redeem_funds',
-		});
-
-		const withdrawn = tx.moveCall({
-			target: '0x2::balance::withdraw_all',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [balResult],
-		});
-
-		tx.moveCall({
-			target: '0x2::balance::send_funds',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [withdrawn, tx.pure.address('0x456')],
-		});
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow?.amount).toBe(500000000n);
-	});
-
-	it('handles coin::divide_into_n (assumes full consumption)', async () => {
-		const client = new MockSuiClient();
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		const coin = tx.object(TEST_COIN_1_ID); // 5 SUI
-
-		tx.moveCall({
-			target: '0x2::coin::divide_into_n',
-			typeArguments: ['0x2::sui::SUI'],
-			arguments: [coin, tx.pure.u64(3n)],
-		});
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		const suiFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === SUI);
-		expect(suiFlow?.amount).toBe(5010000000n); // full coin + gas budget
-	});
-
-	// --- Join operations via MoveCall ---
-
-	it('handles coin::join via MoveCall', async () => {
-		const client = new MockSuiClient();
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		const coin1 = tx.object(TEST_COIN_1_ID); // 5 SUI
-		const coin2 = tx.object(TEST_COIN_2_ID); // 2.5 SUI
-
-		tx.moveCall({
-			target: '0x2::coin::join',
-			typeArguments: ['0x2::sui::SUI'],
-			arguments: [coin1, coin2],
-		});
-
-		tx.transferObjects([coin1], tx.pure.address('0x456'));
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		const suiFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === SUI);
-		expect(suiFlow?.amount).toBe(7510000000n); // 5 + 2.5 SUI + gas
-	});
-
-	it('handles balance::join via MoveCall', async () => {
-		const client = new MockSuiClient();
-		const { tx, result: bal1 } = buildWithdrawalTx({
-			withdrawalAmount: 300000000n,
-			coinType: '0xa0b::usdc::USDC',
-			redeemTarget: '0x2::balance::redeem_funds',
-		});
-
-		// Create a second balance from another withdrawal
-		const withdrawal2 = tx.withdrawal({ amount: 200000000n, type: '0xa0b::usdc::USDC' });
-
-		const bal2 = tx.moveCall({
-			target: '0x2::balance::redeem_funds',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [withdrawal2],
-		});
-
-		// Join bal2 into bal1
-		tx.moveCall({
-			target: '0x2::balance::join',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [bal1, bal2],
-		});
-
-		// Send joined balance to other
-		tx.moveCall({
-			target: '0x2::balance::send_funds',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [bal1, tx.pure.address('0x456')],
-		});
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow?.amount).toBe(500000000n); // 300M + 200M
-	});
-
-	it('handles coin::put (Coin → Balance merge)', async () => {
-		const client = new MockSuiClient();
-		const { tx, result: balResult } = buildWithdrawalTx({
-			withdrawalAmount: 300000000n,
-			coinType: '0xa0b::usdc::USDC',
-			redeemTarget: '0x2::balance::redeem_funds',
-		});
-
-		// Take 100M from balance as a coin
-		const takenCoin = tx.moveCall({
-			target: '0x2::coin::take',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [balResult, tx.pure.u64(100000000n)],
-		});
-
-		// Put it back
-		tx.moveCall({
-			target: '0x2::coin::put',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [balResult, takenCoin],
-		});
-
-		// Send everything back to sender
-		tx.moveCall({
-			target: '0x2::balance::send_funds',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [balResult, tx.pure.address(DEFAULT_SENDER)],
-		});
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow?.amount).toBe(0n); // Everything returned
-	});
-
-	// --- Zero operations ---
-
-	it('handles coin::zero and balance::zero', async () => {
-		const client = new MockSuiClient();
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		tx.moveCall({
-			target: '0x2::coin::zero',
-			typeArguments: ['0x2::sui::SUI'],
-			arguments: [],
-		});
-
-		tx.moveCall({
-			target: '0x2::balance::zero',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [],
-		});
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		const suiFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === SUI);
-		expect(suiFlow?.amount).toBe(10000000n); // gas budget only
-
-		// No USDC outflow
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow).toBeUndefined();
-	});
-
-	it('handles coin::destroy_zero and balance::destroy_zero', async () => {
-		const client = new MockSuiClient();
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		const zeroCoin = tx.moveCall({
-			target: '0x2::coin::zero',
-			typeArguments: ['0x2::sui::SUI'],
-			arguments: [],
-		});
-
-		tx.moveCall({
-			target: '0x2::coin::destroy_zero',
-			typeArguments: ['0x2::sui::SUI'],
-			arguments: [zeroCoin],
-		});
-
-		const zeroBal = tx.moveCall({
-			target: '0x2::balance::zero',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [],
-		});
-
-		tx.moveCall({
-			target: '0x2::balance::destroy_zero',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [zeroBal],
-		});
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		const suiFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === SUI);
-		expect(suiFlow?.amount).toBe(10000000n); // gas budget only
-	});
-});
-
-describe('Coin Flows - Combination Chain Tests', () => {
-	it('redeem → from_balance → split → transfer + send_funds remainder', async () => {
-		const client = new MockSuiClient();
-		const { tx, result: balResult } = buildWithdrawalTx({
-			withdrawalAmount: 200000000n,
-			coinType: '0xa0b::usdc::USDC',
-			redeemTarget: '0x2::balance::redeem_funds',
-		});
-
-		// Convert to coin
-		const coin = tx.moveCall({
-			target: '0x2::coin::from_balance',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [balResult],
-		});
-
-		// Split 150M out
-		const splitResult = tx.moveCall({
-			target: '0x2::coin::split',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [coin, tx.pure.u64(150000000n)],
-		});
-
-		// Transfer split to other
-		tx.transferObjects([splitResult], tx.pure.address('0x456'));
-
-		// Send remainder (50M) back to sender
-		tx.moveCall({
-			target: '0x2::coin::send_funds',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [coin, tx.pure.address(DEFAULT_SENDER)],
-		});
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow?.amount).toBe(150000000n);
-	});
-
-	it('multiple redeems same type', async () => {
-		const client = new MockSuiClient();
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		// First redeem
-		const w1 = tx.withdrawal({ amount: 100000000n, type: '0xa0b::usdc::USDC' });
-		const bal1 = tx.moveCall({
-			target: '0x2::balance::redeem_funds',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [w1],
-		});
-
-		// Second redeem
-		const w2 = tx.withdrawal({ amount: 100000000n, type: '0xa0b::usdc::USDC' });
-		const bal2 = tx.moveCall({
-			target: '0x2::balance::redeem_funds',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [w2],
-		});
-
-		// Consume both
-		tx.moveCall({
-			target: '0x999::test::consume_coin',
-			typeArguments: [normalizeStructTag('0xa0b::usdc::USDC')],
-			arguments: [bal1, bal2],
-		});
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow?.amount).toBe(200000000n);
-	});
-
-	it('redeem + owned coin merge → split → into_balance → consume', async () => {
-		const client = new MockSuiClient();
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		// Redeem 100M USDC from AB
-		const w = tx.withdrawal({ amount: 100000000n, type: '0xa0b::usdc::USDC' });
-		const redeemCoin = tx.moveCall({
-			target: '0x2::coin::redeem_funds',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [w],
-		});
-
-		// Owned USDC coin (500M from mockData)
-		const ownedCoin = tx.object(TEST_USDC_COIN_ID);
-
-		// Join redeem into owned
-		tx.moveCall({
-			target: '0x2::coin::join',
-			typeArguments: [normalizeStructTag('0xa0b::usdc::USDC')],
-			arguments: [ownedCoin, redeemCoin],
-		});
-
-		// Split 250M
-		const [splitCoin] = tx.splitCoins(ownedCoin, [250000000n]);
-
-		// Convert to balance
-		const balance = tx.moveCall({
-			target: '0x2::coin::into_balance',
-			typeArguments: [normalizeStructTag('0xa0b::usdc::USDC')],
-			arguments: [splitCoin],
-		});
-
-		// Consume
-		tx.moveCall({
-			target: '0x999::test::consume_coin',
-			typeArguments: [normalizeStructTag('0xa0b::usdc::USDC')],
-			arguments: [balance, balance],
-		});
-
-		// Send remainder back to sender
-		tx.moveCall({
-			target: '0x2::coin::send_funds',
-			typeArguments: [normalizeStructTag('0xa0b::usdc::USDC')],
-			arguments: [ownedCoin, tx.pure.address(DEFAULT_SENDER)],
-		});
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		// 500M owned + 100M redeemed = 600M. Split 250M. Remainder 350M returned.
-		// Outflow: 600M (raw) - 350M (returned) = 250M
-		expect(usdcFlow?.amount).toBe(250000000n);
-	});
-
-	it('into_balance → balance::join → balance::send_funds to self (no outflow)', async () => {
-		const client = new MockSuiClient();
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		const coin1 = tx.object(TEST_COIN_1_ID); // 5 SUI
-		const coin2 = tx.object(TEST_COIN_2_ID); // 2.5 SUI
-
-		// Convert both to balances
-		const bal1 = tx.moveCall({
-			target: '0x2::coin::into_balance',
-			typeArguments: ['0x2::sui::SUI'],
-			arguments: [coin1],
-		});
-
-		const bal2 = tx.moveCall({
-			target: '0x2::coin::into_balance',
-			typeArguments: ['0x2::sui::SUI'],
-			arguments: [coin2],
-		});
-
-		// Join bal2 into bal1
-		tx.moveCall({
-			target: '0x2::balance::join',
-			typeArguments: ['0x2::sui::SUI'],
-			arguments: [bal1, bal2],
-		});
-
-		// Send back to sender
-		tx.moveCall({
-			target: '0x2::balance::send_funds',
-			typeArguments: ['0x2::sui::SUI'],
-			arguments: [bal1, tx.pure.address(DEFAULT_SENDER)],
-		});
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		const suiFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === SUI);
-		expect(suiFlow?.amount).toBe(10000000n); // gas budget only
-	});
-
-	it('mixed coin types with AB withdrawal', async () => {
-		const client = new MockSuiClient();
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		// USDC from AB withdrawal
-		const w = tx.withdrawal({ amount: 100000000n, type: '0xa0b::usdc::USDC' });
-		const usdcBal = tx.moveCall({
-			target: '0x2::coin::redeem_funds',
-			typeArguments: ['0xa0b::usdc::USDC'],
-			arguments: [w],
-		});
-
-		// Transfer USDC to other
-		tx.transferObjects([usdcBal], tx.pure.address('0x456'));
-
-		// SUI from gas coin, transfer to self
-		const [suiSplit] = tx.splitCoins(tx.gas, [50000000n]);
-		tx.transferObjects([suiSplit], tx.pure.address(DEFAULT_SENDER));
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow?.amount).toBe(100000000n);
-
-		const suiFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === SUI);
-		expect(suiFlow?.amount).toBe(10000000n); // gas budget only, SUI returned to self
-	});
-});
-
-describe('Coin Flows - tx.coin() / tx.balance() Integration', () => {
-	// Helper to resolve intents and analyze in one step.
-	// toJSON({ client }) runs prepareForSerialization which resolves CoinWithBalance intents.
-	async function analyzeWithIntents(tx: Transaction, client: MockSuiClient) {
-		const json = await tx.toJSON({ client });
-		return analyze({ coinFlows }, { client, transaction: json });
-	}
-
-	// --- tx.coin() with SUI ---
-
-	it('tx.coin() SUI from gas coin (coins only, no AB)', async () => {
-		const client = new MockSuiClient();
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		const coin = tx.coin({ balance: 1_000_000_000n }); // 1 SUI from gas
-		tx.transferObjects([coin], tx.pure.address('0x456'));
-
-		const results = await analyzeWithIntents(tx, client);
-
-		const suiFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === SUI);
-		// 1 SUI transferred + gas budget
-		expect(suiFlow).toBeDefined();
-		expect(suiFlow!.amount).toBeGreaterThanOrEqual(1_000_000_000n);
-	});
-
-	it('tx.coin() SUI from gas coin, transfer to self (no outflow beyond gas)', async () => {
-		const client = new MockSuiClient();
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		const coin = tx.coin({ balance: 500_000_000n });
-		tx.transferObjects([coin], tx.pure.address(DEFAULT_SENDER));
-
-		const results = await analyzeWithIntents(tx, client);
-
-		const suiFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === SUI);
-		expect(suiFlow).toBeDefined();
-		// Only gas budget should be the outflow (split returned to self)
-		expect(suiFlow!.amount).toBeLessThan(500_000_000n);
-	});
-
-	// --- tx.coin() with non-SUI ---
-
-	it('tx.coin() non-SUI from coins only (no AB)', async () => {
-		const client = new MockSuiClient();
-		// Default mock has 1000 USDC in coins, 0 AB
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		const coin = tx.coin({ type: '0xa0b::usdc::USDC', balance: 200_000_000n });
-		tx.transferObjects([coin], tx.pure.address('0x456'));
-
-		const results = await analyzeWithIntents(tx, client);
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow).toBeDefined();
-		expect(usdcFlow!.amount).toBe(200_000_000n);
-	});
-
-	it('tx.coin() non-SUI from AB only (no coins)', async () => {
-		// Use a token type that has no default coin objects, only AB
-		const freshClient = new MockSuiClient();
-		freshClient.setAddressBalance(DEFAULT_SENDER, '0xa0b::usdc::USDC', 500_000_000n);
-		// Remove default USDC coins by creating a client that only has SUI coins
-		// Actually the default mock already has USDC coins. Let's test with a type that has no coins.
-		freshClient.setAddressBalance(DEFAULT_SENDER, '0xfff::token::TOKEN', 500_000_000n);
-
-		// Register the token type function for the moveFunctions analyzer
-		freshClient.addMoveFunction({
-			packageId: '0x999',
-			moduleName: 'test',
-			name: 'use_token',
-			visibility: 'public',
-			isEntry: false,
-			parameters: [
-				{
-					reference: null,
-					body: {
-						$kind: 'datatype',
-						datatype: {
-							typeName: '0x2::coin::Coin',
-							typeParameters: [{ $kind: 'typeParameter', index: 0 }],
-						},
-					},
-				},
-			],
-		});
-
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		const coin = tx.coin({ type: '0xfff::token::TOKEN', balance: 200_000_000n, useGasCoin: false });
-		tx.transferObjects([coin], tx.pure.address('0x456'));
-
-		const results = await analyzeWithIntents(tx, freshClient);
-
-		const tokenType = normalizeStructTag('0xfff::token::TOKEN');
-		const tokenFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === tokenType);
-		expect(tokenFlow).toBeDefined();
-		expect(tokenFlow!.amount).toBe(200_000_000n);
-	});
-
-	it('tx.coin() non-SUI from coins + AB combined', async () => {
-		const client = new MockSuiClient();
-		// Default mock has 1000 USDC in coins. Add 500 USDC AB.
-		client.setAddressBalance(DEFAULT_SENDER, '0xa0b::usdc::USDC', 500_000_000n);
-
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		// Request more than coins alone can provide (1000 USDC coins + 500 AB = 1500 total)
-		const coin = tx.coin({ type: '0xa0b::usdc::USDC', balance: 1_200_000_000n });
-		tx.transferObjects([coin], tx.pure.address('0x456'));
-
-		const results = await analyzeWithIntents(tx, client);
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow).toBeDefined();
-		expect(usdcFlow!.amount).toBe(1_200_000_000n);
-	});
-
-	// --- tx.balance() with SUI ---
-
-	it('tx.balance() SUI from gas coin', async () => {
-		const client = new MockSuiClient();
-		client.setAddressBalance(DEFAULT_SENDER, '0x2::sui::SUI', 2_000_000_000n);
-
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		const balance = tx.balance({ balance: 500_000_000n });
-		// Use the balance in a MoveCall
-		tx.moveCall({
-			target: '0x999::test::consume_coin',
-			typeArguments: [normalizeStructTag('0x2::sui::SUI')],
-			arguments: [balance, balance],
-		});
-
-		const results = await analyzeWithIntents(tx, client);
-
-		const suiFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === SUI);
-		expect(suiFlow).toBeDefined();
-		expect(suiFlow!.amount).toBeGreaterThanOrEqual(500_000_000n);
-	});
-
-	// --- tx.balance() with non-SUI ---
-
-	it('tx.balance() non-SUI from AB only (Path 1 - direct withdrawal)', async () => {
-		const client = new MockSuiClient();
-		// Set up enough AB so Path 1 (all-balance + AB sufficient) is used
-		client.setAddressBalance(DEFAULT_SENDER, '0xa0b::usdc::USDC', 1_000_000_000n);
-
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		const balance = tx.balance({ type: '0xa0b::usdc::USDC', balance: 300_000_000n });
-		tx.moveCall({
-			target: '0x999::test::consume_coin',
-			typeArguments: [normalizeStructTag('0xa0b::usdc::USDC')],
-			arguments: [balance, balance],
-		});
-
-		const results = await analyzeWithIntents(tx, client);
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow).toBeDefined();
-		expect(usdcFlow!.amount).toBe(300_000_000n);
-	});
-
-	it('tx.balance() non-SUI from coins only (Path 2 - merge and split)', async () => {
-		const client = new MockSuiClient();
-		// Default: 1000 USDC in coins, 0 AB → Path 2
-
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		const balance = tx.balance({ type: '0xa0b::usdc::USDC', balance: 300_000_000n });
-		tx.moveCall({
-			target: '0x999::test::consume_coin',
-			typeArguments: [normalizeStructTag('0xa0b::usdc::USDC')],
-			arguments: [balance, balance],
-		});
-
-		const results = await analyzeWithIntents(tx, client);
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow).toBeDefined();
-		expect(usdcFlow!.amount).toBe(300_000_000n);
-	});
-
-	it('tx.balance() non-SUI from coins + AB (Path 2 with AB top-up)', async () => {
-		const client = new MockSuiClient();
-		// 1000 USDC in coins + 500 AB. Request more than coins alone.
-		client.setAddressBalance(DEFAULT_SENDER, '0xa0b::usdc::USDC', 500_000_000n);
-
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		const balance = tx.balance({ type: '0xa0b::usdc::USDC', balance: 1_200_000_000n });
-		tx.moveCall({
-			target: '0x999::test::consume_coin',
-			typeArguments: [normalizeStructTag('0xa0b::usdc::USDC')],
-			arguments: [balance, balance],
-		});
-
-		const results = await analyzeWithIntents(tx, client);
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow).toBeDefined();
-		expect(usdcFlow!.amount).toBe(1_200_000_000n);
-	});
-
-	// --- Mixed tx.coin() + tx.balance() ---
-
-	it('tx.coin() + tx.balance() same non-SUI type', async () => {
-		const client = new MockSuiClient();
-		client.setAddressBalance(DEFAULT_SENDER, '0xa0b::usdc::USDC', 500_000_000n);
-
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		const coin = tx.coin({ type: '0xa0b::usdc::USDC', balance: 100_000_000n });
-		const balance = tx.balance({ type: '0xa0b::usdc::USDC', balance: 200_000_000n });
-
-		tx.transferObjects([coin], tx.pure.address('0x456'));
-		tx.moveCall({
-			target: '0x999::test::consume_coin',
-			typeArguments: [normalizeStructTag('0xa0b::usdc::USDC')],
-			arguments: [balance, balance],
-		});
-
-		const results = await analyzeWithIntents(tx, client);
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow).toBeDefined();
-		expect(usdcFlow!.amount).toBe(300_000_000n); // 100 + 200
-	});
-
-	it('tx.coin() SUI + tx.balance() non-SUI in same transaction', async () => {
-		const client = new MockSuiClient();
-		client.setAddressBalance(DEFAULT_SENDER, '0xa0b::usdc::USDC', 500_000_000n);
-
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		const suiCoin = tx.coin({ balance: 1_000_000_000n });
-		const usdcBalance = tx.balance({ type: '0xa0b::usdc::USDC', balance: 200_000_000n });
-
-		tx.transferObjects([suiCoin], tx.pure.address('0x456'));
-		tx.moveCall({
-			target: '0x999::test::consume_coin',
-			typeArguments: [normalizeStructTag('0xa0b::usdc::USDC')],
-			arguments: [usdcBalance, usdcBalance],
-		});
-
-		const results = await analyzeWithIntents(tx, client);
-
-		const suiFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === SUI);
-		expect(suiFlow).toBeDefined();
-		expect(suiFlow!.amount).toBeGreaterThanOrEqual(1_000_000_000n);
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow).toBeDefined();
-		expect(usdcFlow!.amount).toBe(200_000_000n);
-	});
-
-	it('multiple tx.coin() same non-SUI type', async () => {
-		const client = new MockSuiClient();
-
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		// 3 coins of 100 USDC each (default has 1000 USDC total)
-		const coin1 = tx.coin({ type: '0xa0b::usdc::USDC', balance: 100_000_000n });
-		const coin2 = tx.coin({ type: '0xa0b::usdc::USDC', balance: 100_000_000n });
-		const coin3 = tx.coin({ type: '0xa0b::usdc::USDC', balance: 100_000_000n });
-
-		tx.transferObjects([coin1, coin2, coin3], tx.pure.address('0x456'));
-
-		const results = await analyzeWithIntents(tx, client);
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow).toBeDefined();
-		expect(usdcFlow!.amount).toBe(300_000_000n);
-	});
-
-	it('multiple tx.balance() same non-SUI type from AB (Path 1)', async () => {
-		const client = new MockSuiClient();
-		client.setAddressBalance(DEFAULT_SENDER, '0xa0b::usdc::USDC', 1_000_000_000n);
-
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		const b1 = tx.balance({ type: '0xa0b::usdc::USDC', balance: 100_000_000n });
-		const b2 = tx.balance({ type: '0xa0b::usdc::USDC', balance: 150_000_000n });
-
-		tx.moveCall({
-			target: '0x999::test::consume_coin',
-			typeArguments: [normalizeStructTag('0xa0b::usdc::USDC')],
-			arguments: [b1, b2],
-		});
-
-		const results = await analyzeWithIntents(tx, client);
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow).toBeDefined();
-		expect(usdcFlow!.amount).toBe(250_000_000n);
-	});
-
-	// --- Zero balance ---
-
-	it('tx.coin() with zero balance', async () => {
-		const client = new MockSuiClient();
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		const coin = tx.coin({ type: '0xa0b::usdc::USDC', balance: 0n });
-		tx.transferObjects([coin], tx.pure.address('0x456'));
-
-		const results = await analyzeWithIntents(tx, client);
-
-		// No USDC outflow for zero balance
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow).toBeUndefined();
-	});
-
-	it('tx.balance() with zero balance', async () => {
-		const client = new MockSuiClient();
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		const balance = tx.balance({ type: '0xa0b::usdc::USDC', balance: 0n });
-		tx.moveCall({
-			target: '0x999::test::consume_coin',
-			typeArguments: [normalizeStructTag('0xa0b::usdc::USDC')],
-			arguments: [balance, balance],
-		});
-
-		const results = await analyzeWithIntents(tx, client);
-
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow).toBeUndefined();
-	});
-});
-
-describe('Coin Flows - Transfer to Self Bug Fix', () => {
-	it('should not create outflow for non-SUI transfer to self', async () => {
-		const client = new MockSuiClient();
-		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
-
-		const usdcCoin = tx.object(TEST_USDC_COIN_ID); // 500M USDC
-		const [splitCoin] = tx.splitCoins(usdcCoin, [200000000n]);
-
-		// Transfer USDC back to sender
-		tx.transferObjects([splitCoin], tx.pure.address(DEFAULT_SENDER));
-
-		const results = await analyze({ coinFlows }, { client, transaction: await tx.toJSON() });
-
-		// USDC should have zero outflow (transferred to self)
-		const usdcFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === USDC);
-		expect(usdcFlow?.amount).toBe(0n);
-
-		// SUI should have gas budget outflow only
-		const suiFlow = results.coinFlows.result?.outflows.find((f) => f.coinType === SUI);
-		expect(suiFlow?.amount).toBe(10000000n);
 	});
 });
